[feature/menuList] 코드 중간 정리
