+++ src/components/index.js
@@ -1,2 +1,2 @@
-export * from './Button'
-export * from './Input'
+export * from './Button';
+export * from './Input';
