+++ src/features/detail/index.js
@@ -1,2 +1,2 @@
-export * from './OrderDetail'
-export * from './OrderList'
+export * from './OrderDetail';
+export * from './OrderList';
