+++ src/features/order/index.js
@@ -1,1 +1,1 @@
-export * from './MenuList'
+export * from './MenuList';
