+++ src/layout/Header.js
@@ -1,6 +1,6 @@
-import styled from '@emotion/styled'
-import React from 'react'
-import { Link } from 'react-router-dom'
+import styled from '@emotion/styled';
+import React from 'react';
+import { Link } from 'react-router-dom';
 
 export function Header() {
   return (
@@ -8,8 +8,8 @@ export function Header() {
       <Link to={'/order'}>order</Link>
       <Link to={'/detail'}>detail</Link>
     </HeaderStyled>
-  )
-}
+  );
+};
 
 const HeaderStyled = styled.div`
-`
+`;
