+++ src/routes/index.js
@@ -1,1 +1,1 @@
-export * from './Routes'
+export * from './Routes';
